backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pydantic==2.8.2
firebase-admin==6.5.0
openai==1.47.0
requests==2.32.3
